Restrict which types are allowed to be embedded in xml.
This commit changes the "Implicitly converted X to it's string, representation"
runtime warnings into a compiler error. A clean compiler error message should
appear when invalid types are inserted into xml literals.

This change eliminates the need for implicit conversions to `xml.Node` in the
mhtml package object.

The scala-xml-light project is merged into the monadic-html repo, in order to
include mhtml.Rx as part of the xml api.